Restructure test/ into proper project layout
Consolidate test_native/ and test_hardware/ into a single test/directory
with lib/ subdirectories for infrastructure headers, shorter test file
names (dropped at_handler_ prefix), and a unified platformio.ini.
