+++ cmake/utils.cmake
@@ -3,51 +3,25 @@ if(ASYNCAT_HANDLER_BUILD_TESTS)
   add_compile_definitions(LOG_LEVEL=${LOG_LEVEL})
   enable_testing()
 
-  # Only get test files from the TEST_DIR
+  # Only glob test_*.cpp to exclude main.cpp (ESP32-only entry point)
   file(GLOB TEST_FILES CONFIGURE_DEPENDS
-    ${TEST_DIR}/test_native/*.cpp
+    ${TEST_DIR}/test/test_*.cpp
   )
 
   foreach(TEST_FILE ${TEST_FILES})
     get_filename_component(TEST_NAME ${TEST_FILE} NAME_WE)
     set(EXEC_NAME "${TEST_NAME}_test_exec")
 
-    set(TEST_SRC ${TEST_FILE})
+    add_executable(${EXEC_NAME} ${TEST_FILE})
 
-    add_executable(${EXEC_NAME} ${TEST_SRC})
-
-    target_link_libraries(${EXEC_NAME}
+    target_include_directories(${EXEC_NAME}
       PRIVATE
-      ArduinoNativeMocks
-      AsyncATHandler
-      gmock
-      gtest
-      gtest_main
-      unity
+      ${TEST_DIR}
+      ${TEST_DIR}/lib/BoneBuilder
+      ${TEST_DIR}/lib/Mocks
+      ${TEST_DIR}/lib/SerialCommunicator
+      ${TEST_DIR}/lib/TestCommon
     )
-
-    add_test(NAME ${EXEC_NAME} COMMAND ${EXEC_NAME})
-  endforeach()
-endif()
-
-
-# === HYBRID TESTS ON NATIVE ===
-if(ASYNCAT_HANDLER_BUILD_TESTS)
-  add_compile_definitions(LOG_LEVEL=${LOG_LEVEL})
-  enable_testing()
-
-  # Only get test files from the TEST_DIR
-  file(GLOB TEST_FILES CONFIGURE_DEPENDS
-    ${TEST_DIR}/test_hybrid/*.cpp
-  )
-
-  foreach(TEST_FILE ${TEST_FILES})
-    get_filename_component(TEST_NAME ${TEST_FILE} NAME_WE)
-    set(EXEC_NAME "${TEST_NAME}_htest_exec")
-
-    set(TEST_SRC ${TEST_FILE})
-
-    add_executable(${EXEC_NAME} ${TEST_SRC})
 
     target_link_libraries(${EXEC_NAME}
       PRIVATE
