+++ test/platformio.ini
@@ -1,0 +1,19 @@
+[env]
+platform = espressif32
+board = esp32dev
+framework = arduino
+test_framework = googletest
+test_dir = test/
+
+build_flags =
+  -std=gnu++17
+  -D CORE_DEBUG_LEVEL=5
+
+build_unflags =
+  -std=gnu++11
+
+[env:test]
+lib_deps = ..
+
+[env:ci]
+lib_deps = ..
