+++ test/test/test_http.cpp
@@ -32,7 +32,7 @@ class AsyncATHandlerHTTPTest : public FreeRTOSTest {
       }
       bool success = CleanupATHandler(handler);
       if (!success) { log_w("Handler teardown may have failed"); }
-      std::this_thread::sleep_for(std::chrono::milliseconds(200));
+      delay(200);
       delete handler;
       handler = nullptr;
     }
