--- test/test_hardware/.gitignore
@@ -1,1 +1,0 @@
-.pio
