--- test/test_hardware/lib/Peripherals/Peripherals.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "Expanders/Expanders.h"
-#include "Mux/Mux.h"
-#include "UartMux/UartMux.h"
-
-void startPeripherals();
